requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opencv-python==4.10.0.84
numpy
ultralytics==8.3.0
pydantic==2.8.2
requests
python-dotenv
